# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "National Exam Card API",
        "description": "The National Exam Card API provides access to student's application for exam card data and process. ",
        "version": "0.1.0"
    },
    "paths": {
...
        "/school": {
            "get": {
                "tags": [
                    "school"
                ],
                "summary": "Get a list of all schools",
                "operationId": "a797b139084f697bd0ff6e06407c9857",
                "responses": {
                    "200": {
                        "description": "List of schools"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "school"
                ],
                "summary": "Create a new school",
                "operationId": "38dc7d795b63ce7751a0e25701905827",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Sample content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "School created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/school/{id}": {
            "get": {
                "tags": [
                    "school"
                ],
                "summary": "Get a specific school by ID",
                "operationId": "57b0e5960851cb1483a8e4cfbbecdaa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the school",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "School details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "school"
                ],
                "summary": "Update an existing school",
                "operationId": "4c5044fc02583f859c77bc752a373de9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the school",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "School updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "school"
                ],
                "summary": "Delete a school by ID",
                "operationId": "51cc4ecdaca2e43862fb6aeead1516f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the school",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "School deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}